refactor[Issue] : Create new package for issue
