--- src/main/java/com/keepyuppy/KeepyUppy/content/communication/request/CreateScheduleRequest.java
@@ -1,18 +1,0 @@
-package com.keepyuppy.KeepyUppy.content.communication.request;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Data
-@NoArgsConstructor
-@Schema(name = "스케줄 생성 요청")
-public class CreateScheduleRequest {
-
-    private String title;
-    private String content;
-    private LocalDateTime startDateTime;
-    private LocalDateTime endDateTime;
-}
