--- src/main/java/com/keepyuppy/KeepyUppy/content/domain/entity/IssueAssignment.java
@@ -1,46 +1,0 @@
-package com.keepyuppy.KeepyUppy.content.domain.entity;
-
-import com.keepyuppy.KeepyUppy.member.domain.entity.Member;
-import jakarta.persistence.*;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.Objects;
-
-@Entity
-@Getter
-@NoArgsConstructor
-public class IssueAssignment {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "issue_id")
-    private Issue issue;
-
-    @ManyToOne
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    @Builder
-    public IssueAssignment(Issue issue, Member member) {
-        this.issue = issue;
-        this.member = member;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        IssueAssignment that = (IssueAssignment) o;
-        return Objects.equals(id, that.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
-}
