--- src/main/java/com/keepyuppy/KeepyUppy/content/domain/enums/ContentType.java
@@ -1,8 +1,0 @@
-package com.keepyuppy.KeepyUppy.content.domain.enums;
-
-public enum ContentType {
-    POST,
-    ANNOUNCEMENT,
-    ISSUE,
-    SCHEDULE
-}
