--- src/main/java/com/keepyuppy/KeepyUppy/content/domain/enums/IssueStatus.java
@@ -1,7 +1,0 @@
-package com.keepyuppy.KeepyUppy.content.domain.enums;
-
-public enum IssueStatus {
-    TODO,
-    INPROGRESS,
-    DONE
-}
