--- src/main/java/com/keepyuppy/KeepyUppy/content/repository/PostJpaRepository.java
@@ -1,18 +1,0 @@
-package com.keepyuppy.KeepyUppy.content.repository;
-
-import com.keepyuppy.KeepyUppy.content.domain.entity.Post;
-import com.keepyuppy.KeepyUppy.content.domain.enums.ContentType;
-import com.keepyuppy.KeepyUppy.team.domain.entity.Team;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface PostJpaRepository extends JpaRepository<Post,Long>{
-
-    Page<Post> findByTeamAndTypeOrderByCreatedDateDesc(Team team, ContentType type, Pageable pageable);
-
-}
