--- src/main/java/com/keepyuppy/KeepyUppy/content/repository/PostRepositoryImpl.java
@@ -1,12 +1,0 @@
-package com.keepyuppy.KeepyUppy.content.repository;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-@Repository
-@RequiredArgsConstructor
-public class PostRepositoryImpl {
-
-    private final JPAQueryFactory jpaQueryFactory;
-}
