--- src/main/java/com/keepyuppy/KeepyUppy/content/repository/ScheduleJpaRepository.java
@@ -1,9 +1,0 @@
-package com.keepyuppy.KeepyUppy.content.repository;
-
-import com.keepyuppy.KeepyUppy.content.domain.entity.Schedule;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ScheduleJpaRepository extends JpaRepository<Schedule,Long>{
-}
