+++ src/main/java/com/keepyuppy/KeepyUppy/issue/domain/enums/IssueStatus.java
@@ -1,0 +1,7 @@
+package com.keepyuppy.KeepyUppy.issue.domain.enums;
+
+public enum IssueStatus {
+    TODO,
+    INPROGRESS,
+    DONE
+}
