+++ src/main/java/com/keepyuppy/KeepyUppy/post/communication/response/PostResponse.java
@@ -1,0 +1,30 @@
+package com.keepyuppy.KeepyUppy.post.communication.response;
+
+import com.keepyuppy.KeepyUppy.post.domain.entity.Post;
+import com.keepyuppy.KeepyUppy.post.domain.enums.ContentType;
+import com.keepyuppy.KeepyUppy.member.communication.response.MemberResponse;
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@Schema(name = "게시글, 공지 정보 응답")
+@AllArgsConstructor
+public class PostResponse {
+
+    private Long id;
+    private String title;
+    private MemberResponse author;
+    private String content;
+    private Boolean isAnnouncement;
+
+    public static PostResponse of(Post post){
+        return new PostResponse(
+                post.getId(),
+                post.getTitle(),
+                new MemberResponse(post.getAuthor()),
+                post.getContent(),
+                post.getType() == ContentType.ANNOUNCEMENT
+        );
+    }
+}
