+++ src/main/java/com/keepyuppy/KeepyUppy/post/domain/enums/ContentType.java
@@ -1,0 +1,8 @@
+package com.keepyuppy.KeepyUppy.post.domain.enums;
+
+public enum ContentType {
+    POST,
+    ANNOUNCEMENT,
+    ISSUE,
+    SCHEDULE
+}
