+++ src/main/java/com/keepyuppy/KeepyUppy/post/repository/PostJpaRepository.java
@@ -1,0 +1,16 @@
+package com.keepyuppy.KeepyUppy.post.repository;
+
+import com.keepyuppy.KeepyUppy.post.domain.entity.Post;
+import com.keepyuppy.KeepyUppy.post.domain.enums.ContentType;
+import com.keepyuppy.KeepyUppy.team.domain.entity.Team;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PostJpaRepository extends JpaRepository<Post,Long>{
+
+    Page<Post> findByTeamAndTypeOrderByCreatedDateDesc(Team team, ContentType type, Pageable pageable);
+
+}
