+++ src/main/java/com/keepyuppy/KeepyUppy/post/repository/PostRepositoryImpl.java
@@ -1,0 +1,12 @@
+package com.keepyuppy.KeepyUppy.post.repository;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+@Repository
+@RequiredArgsConstructor
+public class PostRepositoryImpl {
+
+    private final JPAQueryFactory jpaQueryFactory;
+}
