+++ src/main/java/com/keepyuppy/KeepyUppy/post/repository/ReadAnnouncementRepository.java
@@ -1,0 +1,9 @@
+package com.keepyuppy.KeepyUppy.post.repository;
+
+import com.keepyuppy.KeepyUppy.post.domain.entity.ReadAnnouncement;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ReadAnnouncementRepository extends JpaRepository<ReadAnnouncement,Long> {
+}
