+++ src/main/java/com/keepyuppy/KeepyUppy/post/repository/ScheduleJpaRepository.java
@@ -1,0 +1,9 @@
+package com.keepyuppy.KeepyUppy.post.repository;
+
+import com.keepyuppy.KeepyUppy.post.domain.entity.Schedule;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ScheduleJpaRepository extends JpaRepository<Schedule,Long>{
+}
